requirements.txt
fastapi 
uvicorn[standard] 
asyncpg
sqlalchemy[asyncio]
requests>=2.31.0
python-dotenv>=1.0.0
# google-cloud-functions-framework>=3.4.0
# google-cloud-logging>=3.5.0

# For development/testing
pytest
# pytest>=7.4.0
python-dateutil>=2.8.2
pytest-mock
